+++ src/main/java/com/paytm/inventorymanagement/repositories/InventoryRepository.java
@@ -3,5 +3,5 @@ package com.paytm.inventorymanagement.repositories;
 import com.paytm.inventorymanagement.models.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
 }
